% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzTools-package.R
\docType{package}
\name{AzTools-package}
\alias{AzTools}
\alias{AzTools-package}
\title{AzTools: Tools for Az}
\description{
This package is for tools that I use
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ajmackay/AzTools}
  \item Report bugs at \url{https://github.com/ajmackay/AzTools/issues}
}

}
\author{
\strong{Maintainer}: Aaron Mackay \email{mackay_aaron@protonmail.com}

}
\keyword{internal}
