% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob_utils.R
\name{Blob_Exists}
\alias{Blob_Exists}
\title{Check if a blob exists in certain directory}
\usage{
Blob_Exists(blob_name, directory, account_name, container_name, sas_token)
}
\arguments{
\item{blob_name}{String. Name of the blob.}

\item{directory}{String. Path where the blob will be searched.}

\item{account_name}{String. Storage account name.}

\item{container_name}{String. Container name.}

\item{sas_token}{String. SAS token with the requiered permissions to upload.}
}
\value{
TRUE if the blob exists, FALSE if not.
}
\description{
This function checks if a blob exists in certain directory
}
