% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{create.project}
\alias{create.project}
\title{create.project}
\usage{
create.project(
  project.name,
  alt.dir = NULL,
  shiny = FALSE,
  data.ticket = FALSE,
  default.queries = NULL,
  initial.commit = TRUE
)
}
\arguments{
\item{project.name}{Character of the name of the project}

\item{alt.dir}{Character of a directory to save the project that isn't the default}

\item{shiny}{Logical: Whether to include components for a shiny application in the project}

\item{data.ticket}{logical for whether this project is for a data ticket in the data ticket folder. If true, then a git repository won't be initialised.}

\item{default.queries}{Character vector of default queries to be included in the new project at setup}

\item{initial.commit}{Logical for whether to automatically do the initial git commit after setting up (mainly used for development purposes)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{        Nothing
}\if{html}{\out{</div>}}
}
\description{
Initialises a new project with a folder structure and git repository. Project will be created in the default R directory unless alt.dir is supplied
}
