% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_date_ref}
\alias{create_date_ref}
\title{Create a Date Reference Table}
\usage{
create_date_ref(start, end, by)
}
\arguments{
\item{start}{Start Date}

\item{end}{End Date}

\item{by}{The interval for which the date reference should be created by (determined by seq.Date())}
}
\value{
Tibble
}
\description{
Create a Date Reference Table
}
