% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{create_project}
\alias{create_project}
\title{Create New R Project}
\usage{
create_project(name, dir, initial_commit = TRUE)
}
\arguments{
\item{name}{Name of the Project}

\item{dir}{Directory of the Project}

\item{initial_commit}{Initialise git repository and perform an initial commit}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{          Nothing
}\if{html}{\out{</div>}}
}
\description{
Creates and opens a new R project with the template files in the AzTools package.
Initialises a git repository and performs an initial commit.
}
