% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_blobs.R
\name{Download_Batch}
\alias{Download_Batch}
\title{Download all files from a directory in an Azure Storage Account's container}
\usage{
Download_Batch(
  az_path = "",
  local_path = ".",
  account_name,
  container_name,
  sas_token
)
}
\arguments{
\item{az_path}{String. Path to the directory that will be downloaded. If provided, it must end with "/".}

\item{local_path}{String. Local directory where the files will be stored.}

\item{account_name}{String. Storage account name.}

\item{container_name}{String. Container name.}

\item{sas_token}{String. SAS token with the required permissions to download the files.}
}
\value{
The output of the download process.
}
\description{
This function downloads all the files located at the given az_path recursively
and saves them in the directory specified by local_path.
}
