% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob_utils.R
\name{Download_Blob}
\alias{Download_Blob}
\title{Download a specific blob}
\usage{
Download_Blob(
  az_blob,
  local_file = "./az_download.file",
  account_name,
  container_name,
  sas_token
)
}
\arguments{
\item{az_blob}{String. Path to the blob.}

\item{local_file}{String. Path to the file where the blob will be downloded.}

\item{account_name}{String. Storage account name.}

\item{container_name}{String. Container name.}

\item{sas_token}{String. SAS token with the requiered permissions to upload the files.}
}
\description{
This function downloads the blob specified at the az_blob param
}
