% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{facet.arrange}
\alias{facet.arrange}
\title{Prepare dataframe for ordering categories across facets}
\usage{
facet.arrange(dat, facet, x, y, smallest.left = TRUE)
}
\arguments{
\item{dat}{Dataframe including at least a categorical variable, a count or proportion of the categorical variable and a grouping variable}

\item{facet}{The variable which you want to facet/group by}

\item{x}{Categorical variable}

\item{y}{Count/Proportion of the categorical variable}

\item{smallest.left}{Defaults to ordering the categorical variables so that the smallest count/proportion is on the left within each facet}
}
\value{
Dataframe with categorical variable prepped for faceting
}
\description{
IN DEVELOPMENT. Prepares a dataframe for easy ordering of a categorical variable across groups/facets used within ggplot.
Data must contain summarised counts and/or proportions of a categorical variable and a grouping variable that is used for faceting.
}
