% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\name{flex2xl}
\alias{flex2xl}
\title{Export a flextable into a .xlsx file}
\usage{
flex2xl(table, path)
}
\arguments{
\item{table}{A flextable}

\item{path}{Path to the Excel file to be created}
}
\value{
Returns an .xlsx file based on a flextable
}
\description{
Export a flextable into a .xlsx file
}
\examples{
\dontrun{
ft <- flextable::flextable(head(mtcars))
# color some cells in blue
ft <- flextable::bg(ft, i=ft$body$dataset$disp>200, j=3, bg = "#7ed6df", part = "body")
# color a few cells in yellow
ft <- flextable::bg(ft, i=ft$body$dataset$vs==0, j=8, bg = "#FCEC20", part = "body")
# export your flextable as a .xlsx in the current working directory
flex2xl(ft, filename ="myFlextable", path="path/to/the/excel_file.xlsx")
}
}
