% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{format_date}
\alias{format_date}
\title{Format dates into human readable format}
\usage{
format_date(start, end)
}
\arguments{
\item{start}{Starting date}

\item{end}{End date}
}
\value{
String
}
\description{
Format dates into human readable format
}
