% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\name{ft.prep}
\alias{ft.prep}
\title{Prepare and Output a Flextable}
\usage{
ft.prep(dat, prep.names = TRUE, digits = 1)
}
\arguments{
\item{dat}{Dataframe}

\item{prep.names}{Turn Names into Titles}
}
\value{
flextable
}
\description{
Prepare and Output a Flextable
}
