% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\name{ft.summary}
\alias{ft.summary}
\title{Basic Summary Table}
\usage{
ft.summary(dat, summ.by = NULL, dp = 1, caption = NULL, include.p = FALSE)
}
\arguments{
\item{dat}{A dataframe to summarise in wide format (one column for each variable to summarise)}

\item{summ.by}{Variable to stratify summary table by}

\item{dp}{Number of decimal places}

\item{caption}{Table Caption}

\item{include.p}{Include P value}
}
\value{
A gtsummary table
}
\description{
Basic Summary Table
}
