% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob_utils.R
\name{List_Blobs}
\alias{List_Blobs}
\title{List all the blobs inside a directory}
\usage{
List_Blobs(az_path, account_name, container_name, sas_token)
}
\arguments{
\item{az_path}{String. Path to the directory inside the blob container.}

\item{account_name}{String. Storage account name.}

\item{container_name}{String. Container name.}

\item{sas_token}{String. SAS token with the requiered permissions to upload the files.}
}
\value{
List of all blobs inside the directory
}
\description{
This function list all blobs inside a directory
}
