% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{load.figs}
\alias{load.figs}
\title{Load Figures and Tables .RData}
\usage{
load.figs(file = "objects/figs-tables")
}
\arguments{
\item{file}{Location and name of figures and tables .RData object}
}
\value{
Saves the objects in Global Environment
}
\description{
Load Figures and Tables .RData
}
