% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{load.objects}
\alias{load.objects}
\title{Load all objects}
\usage{
load.objects(dataset = "all-objects")
}
\arguments{
\item{dataset}{String of the name of the dataset to load. Default is all-objects}
}
\value{
Loads all objects saved in all-objects.RData into the environment
}
\description{
Load all objects
}
