% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{period_label}
\alias{period_label}
\title{Date Label}
\usage{
period_label(
  date,
  type = c("day", "week", "fortnight", "month", "quarter", "half", "year", "fy",
    "finmonth", "finquarter"),
  as.factor = TRUE,
  sep = "_"
)
}
\arguments{
\item{date}{Vector of dates to generate labels for}

\item{type}{Character for which type of label to generate. Only the first one will be used. These labels can be combined into any arbitrary order \cr
or combination by joining different components with a '.'. E.g., if you wanted a label in the format MMM-YY, you would do month.year, with sep = '-', or if instead you wanted YY_MMM, \cr
use year.month, with a sep = _. There is no minimum or maximum number of components for the combination (i.e., year.month.day is valid). \cr
report is report period, quarter and half are quarter/half a year, day is day of year.}

\item{as.factor}{Logical for whether to return as an ordered factor (TRUE), or a character(FALSE). Alternatively, a character/factor vector can be passed as the levels to be used for the resulting factor variable. Any values in the data that don't appear in the provided vector will be converted to NA.}

\item{sep}{Separator between the months and years for month.year and report.period}
}
\value{
Ordered factor or character vector of labels for each date
}
\description{
Label Dates
}
