% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitstuff.R
\name{post.issue}
\alias{post.issue}
\title{Post Issue to AzTools Repository}
\usage{
post.issue(title, body)
}
\arguments{
\item{title}{Title of the Issue}

\item{body}{Description of Issue}
}
\description{
Post Issue to AzTools Repository
}
