% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\name{prep.flex}
\alias{prep.flex}
\title{Prepare and Output a Flextable}
\usage{
prep.flex(dat, prep.names = TRUE)
}
\arguments{
\item{dat}{Dataframe}

\item{prep.names}{Turn Names into Titles}
}
\value{
flextable
}
\description{
Prepare and Output a Flextable
}
