% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{prep_flex}
\alias{prep_flex}
\title{Prepare and Output a Flextable}
\usage{
prep_flex(dat, prep.names = TRUE, digits = 1)
}
\arguments{
\item{dat}{Dataframe}

\item{prep.names}{Turn Names into Titles}
}
\value{
flextable
}
\description{
Prepare and Output a Flextable
}
