% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{record_gg}
\alias{record_gg}
\title{record_gg}
\usage{
record_gg(dir = "outputs/imgs", width = 9, dpi = 300)
}
\arguments{
\item{dir}{The directory in which to save plots to (default is outputs/imgs)}

\item{width}{Width in inches of the image (default is 9)}

\item{dpi}{Default 300}

\item{height}{Height is calculated as a function of width. It will always be less than width}
}
\value{
Image in folder
}
\description{
Uses camcorder::gg_record() to save and view ggplots as they would appear when exported.
When record_gg is used, subsequent ggplots will be saved in the specified folder (default 'outputs/imgs') and the plot
will be viewed in the viewer pane. Use stop_gg() to stop record_gg()
}
