% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{save.figs}
\alias{save.figs}
\title{Save Figures and Tables}
\usage{
save.figs(dir = "objects/", name = "figs-tables")
}
\arguments{
\item{dir}{Folder where to save object}

\item{name}{Name of .RData object}
}
\value{
.RData object saved in objects
}
\description{
Save Figures and Tables
}
