% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{save.flex}
\alias{save.flex}
\title{Save Flex Table}
\usage{
save.flex(dat, name, format = c("word"))
}
\arguments{
\item{dat}{Flextable Object}

\item{name}{String: Name of file to be saved}

\item{format}{Format for file saving (default .docx)}
}
\value{
File
}
\description{
Saves a flextable object in the preferred format to the tables folder in the output directory
}
