% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{save.object}
\alias{save.object}
\title{Save a single object}
\usage{
save.object(obj, name)
}
\arguments{
\item{obj}{The object to Save}

\item{name}{File name of the Object}
}
\value{
Saves the object in data/objects/
}
\description{
\code{save.object()} Saves a specified object as a .rds file in data/objects/
}
