% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{save.objects.all}
\alias{save.objects.all}
\title{Save all objects}
\usage{
save.objects.all(dir = NULL)
}
\arguments{
\item{dir}{directory where all objects should be saved (default is data/objects)}
}
\value{
Saves all the objects and shit
}
\description{
Save all objects
}
