% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\name{set.flex.default}
\alias{set.flex.default}
\title{Set Flextable Defaults}
\usage{
set.flex.default(
  font.size = 10,
  font.family = "Calibri",
  padding = 6,
  digits = 1,
  na.str = "-"
)
}
\arguments{
\item{font.size}{Default is 10}

\item{font.family}{Default is Calibri}

\item{padding}{Default is 6}

\item{theme}{Default is vanilla}
}
\value{
Flextable Defaults
}
\description{
Set Flextable Defaults
}
