% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{stop_quietly}
\alias{stop_quietly}
\title{Stop R session without error}
\usage{
stop_quietly()
}
\value{
Invisible, the R session will be stopped.
}
\description{
\code{stop_quietly()} does exactly the same as \code{stop()}, however it does not print anything to the console.
}
\author{
Stuart K. Grange
}
