% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summtbl.R
\name{summ.tbl}
\alias{summ.tbl}
\title{Basic Summary Table}
\usage{
summ.tbl(dat, summ.by = NULL, dp = 1, caption = NULL)
}
\arguments{
\item{dat}{A dataframe to summarise in wide format (one column for each variable to summarise)}

\item{summ.by}{Variable to stratify summary table by}

\item{dp}{Number of decimal places}

\item{caption}{Table Caption}
}
\value{
A gtsummary table
}
\description{
Basic Summary Table
}
