% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-to-word.R
\name{tbl.to.word}
\alias{tbl.to.word}
\title{Save flextable in a word doc with}
\usage{
\method{tbl}{to.word}(tbl.list)
}
\arguments{
\item{dat}{A dataframe to summarise}

\item{summ.by}{Variable to stratify summary table by}

\item{dp}{Number of decimal places}

\item{caption}{Table Caption}
}
\value{
A gtsummary table
}
\description{
Save flextable in a word doc with
}
