% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{theme_clean}
\alias{theme_clean}
\title{Clean Theme}
\usage{
theme_clean(base_size = 12, dark_text = "#1A242F")
}
\arguments{
\item{base_size}{Base Size of Text}

\item{dark_text}{Colour of the Title}
}
\value{
ggplot2 theme
}
\description{
A simple and clean theme for ggplot
}
