% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{theme_solar}
\alias{theme_solar}
\title{Solarised Theme}
\usage{
theme_solar(bg.colour = "#fcf6e4", ...)
}
\arguments{
\item{bg.colour}{Background Colour of the plot}

\item{...}{Additional arguments for ggplot2::theme()}
}
\value{
ggplot theme
}
\description{
Solarised Theme
}
