% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-extract.R
\name{time.extract}
\alias{time.extract}
\title{Extract Time Units}
\usage{
\method{time}{extract}(data, date_var, time_units)
}
\arguments{
\item{data}{Dataframe with at least one datetime variable}

\item{date_var}{Datetime variable to extract datetime unit from}

\item{time_units}{The units of datetime to extract as a string e.g. c("year", "month", "day")}
}
\value{
A dataframe with specified datetime units
}
\description{
Extract Time Units
}
