% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob_utils.R
\name{Upload_Batch}
\alias{Upload_Batch}
\title{Upload all files from a local directory to an Azure Storage Account's container}
\usage{
Upload_Batch(
  az_path = "",
  local_path = ".",
  account_name,
  container_name,
  sas_token
)
}
\arguments{
\item{az_path}{String. Path to the directory where files will be uploaded. If provided, it must end with "/".}

\item{local_path}{String. Local directory where the files to upload are located.}

\item{account_name}{String. Storage account name.}

\item{container_name}{String. Container name.}

\item{sas_token}{String. SAS token with the required permissions to upload the files.}
}
\value{
The output of the upload process.
}
\description{
This function uploads all files located at the given local_path recursively
and saves them in the directory specified by az_path.
}
