% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob_utils.R
\name{Upload_Blob}
\alias{Upload_Blob}
\title{Upload a specific blob}
\usage{
Upload_Blob(local_file, az_blob, account_name, container_name, sas_token)
}
\arguments{
\item{local_file}{String. Path to the file that will be uploaded}

\item{az_blob}{String. Path to the blob.}

\item{account_name}{String. Storage account name.}

\item{container_name}{String. Container name.}

\item{sas_token}{String. SAS token with the requiered permissions to upload.}
}
\description{
This function uploadss the blob specified at the local_file param
}
