% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{view_gg}
\alias{view_gg}
\title{view_gg}
\usage{
view_gg(plot, width = 9, height = NULL, dpi = 300, ...)
}
\arguments{
\item{plot}{ggplot object}

\item{width}{Width of the plot in inches}

\item{height}{Default is for height to be set as a function of width. This can be overwritten with a manual value}

\item{dpi}{The dpi. Default is 300.}
}
\value{
An Image
}
\description{
Prints a ggplot to the viewer window.
view_gg calls camcorder::gg_record() which prints a ggplot to the viewer window, so that plot measurements can be viewed more accurately.
}
