% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-tools.R
\name{wilson_ci}
\alias{wilson_ci}
\title{Wilson Score Confidence Intervals}
\usage{
wilson_ci(num, denom, ...)
}
\arguments{
\item{num}{Numerator of proportion}

\item{denom}{Denominator of proportion}

\item{...}{Additional arguments passed to prop.test}
}
\value{
Tibble with conf.low and conf.high
}
\description{
Calculates the low and high confidence intervals for proportions using the Wilson Score Interval Method.
}
