% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{load.object}
\alias{load.object}
\title{Load a Single Object}
\usage{
load.object(name)
}
\arguments{
\item{name}{File name of the object}
}
\value{
Loads the object from data/objects
}
\description{
Load a Single Object
}
