% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{save_gg}
\alias{save_gg}
\title{save_gg}
\usage{
save_gg(
  plot,
  file.name = NULL,
  overwrite = TRUE,
  dir = "outputs/imgs/",
  width = 9,
  height = NULL,
  dpi = 300
)
}
\arguments{
\item{plot}{A ggplot}

\item{dpi}{}
}
\description{
Saves a ggplot as a png and prints to the viewer plane
}
