% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{save_gg}
\alias{save_gg}
\title{save_gg}
\usage{
save_gg(
  plot,
  file.name = NULL,
  format = "png",
  overwrite = TRUE,
  dir = "outputs/plots/",
  open = TRUE,
  width = 9,
  height = NULL,
  dpi = 300,
  ...
)
}
\arguments{
\item{plot}{A ggplot}

\item{file.name}{The name of the file without format}

\item{format}{The file format (default .png)}

\item{overwrite}{Overwrite existing file of same file.name (default == TRUE)}

\item{dir}{Directory where to save image. Default output/figs}

\item{width}{Width of plot. Default is 9}

\item{height}{Height of plot. Default is widdth * 9/16}

\item{dpi}{Default 300}
}
\value{
Image file
}
\description{
Saves a ggplot to the specified format and prints to the viewer plane
}
