% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\name{set.default.ft}
\alias{set.default.ft}
\title{Set Flextable Defaults}
\usage{
set.default.ft(
  font.family = "Calibri",
  font.size = 10,
  layout = "autofit",
  padding = 6,
  theme_fun = "theme_vanilla",
  digits = 3,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Set Flextable Defaults
}
